refactor: flatten `indexer/` module into `pagination.rs` and `request.rs`
Move the two submodules directly into `src/`, removing the intermediate
`indexer/` directory.
