+++ dango/testing/src/lib.rs
@@ -5,10 +5,12 @@ mod crypto;
 mod genesis;
 pub mod httpd;
 mod hyperlane;
-mod indexer;
+mod pagination;
 pub mod perps;
+mod request;
 mod setup;
 
 pub use {
-    account::*, account_creation::*, crypto::*, genesis::*, hyperlane::*, indexer::*, setup::*,
+    account::*, account_creation::*, crypto::*, genesis::*, hyperlane::*, pagination::*,
+    request::*, setup::*,
 };
